rewriter/requirements.txt
python-telegram-bot==21.9
openai>=1.50.0
python-dotenv==1.0.0
python-docx
pdfplumber
tavily-python
